Se agrega la lógica de no poder borrar user con posts asociados

+++ Application/Features/Users/Commands/DeleteUser/DeleteUserCommandHandler.cs
@@ -18,6 +18,9 @@ public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, Resul
         var exists = await _unitOfWork.Users.GetByIdAsync(request.UserId, cancellationToken);
         if (exists == null)
             return Result.Failure("El usuario no existe");
+        var existingPosts = await _unitOfWork.Posts.GetByUserIdAsync(request.UserId);
+        if (existingPosts.Count() != 0)
+            return Result.Failure("El usuario aún tiene posts creados");
         await _unitOfWork.Users.DeleteAsync(exists, cancellationToken);
         await _unitOfWork.SaveChangesAsync(cancellationToken);
         return Result.Success();
